Move components into files containing both the jsx and css.
